# services/agent/app/main.py
from contextlib import asynccontextmanager
from datetime import datetime, timedelta
from secrets import compare_digest
from typing import Annotated
from zoneinfo import ZoneInfo

from fastapi import Depends, FastAPI, Header, HTTPException, Request, Response
from fastapi.exceptions import RequestValidationError
from fastapi.responses import JSONResponse
from loguru import logger

from app.config import get_settings
from app.errors import (
    SERVICE_UNAUTHORIZED,
    USER_CONTEXT_REQUIRED,
    VOICE_INVALID_REQUEST,
    PublicError,
)
from app.finance import FinanceStore
from app.models import Workspace
from app.voice.sessions import StartVoice, VoiceConnection, VoiceSessions, VoiceState

# Provider debug logs can include conversation content, so they are normally disabled.
# Temporarily re-enabled while diagnosing the no-user-audio issue: `logger.disable("pipecat")`
# was silencing every pipecat-internal log, including the Daily subscription/media events
# needed to see why the bot never receives microphone audio. Restore the disable() call
# (or replace with a warning-level-only filter) once that is diagnosed.
# logger.disable("pipecat")


@asynccontextmanager
async def lifespan(app: FastAPI):
    settings = get_settings()
    # Importing the complete Pipecat voice graph can take ~20 seconds in a fresh container.
    # Do it before health becomes ready, never on a user's first call.
    logger.info("voice_pipeline_warming engine={}", settings.voice_engine)
    # Cascade and realtime share every tool, prompt, session, and tracing concern;
    # only which pipeline module runs differs, selected by VOICE_ENGINE.
    if settings.voice_engine == "realtime":
        from app.voice.realtime_pipeline import run_realtime as run_pipeline
    else:
        from app.voice.pipeline import run_cascade as run_pipeline

    logger.info("voice_pipeline_warmed")
    app.state.finance = None
    if settings.voice_ready:
        app.state.finance = FinanceStore(settings.database_url)
        await app.state.finance.open()
    app.state.voice = VoiceSessions(
        settings, run_pipeline=run_pipeline, finance_store=app.state.finance
    )
    yield
    await app.state.voice.close()
    if app.state.finance:
        await app.state.finance.close()


app = FastAPI(title="Riverline Agent API", version="0.1.0", lifespan=lifespan)


def _public_detail(detail: object, status_code: int) -> dict[str, str | bool]:
    if isinstance(detail, dict) and isinstance(detail.get("code"), str):
        return detail
    if status_code == 401:
        return SERVICE_UNAUTHORIZED.detail()
    if status_code == 400:
        return USER_CONTEXT_REQUIRED.detail()
    return PublicError(
        "request_failed", "The request could not be completed.", status_code
    ).detail()


@app.exception_handler(HTTPException)
async def http_error_handler(_: Request, error: HTTPException) -> JSONResponse:
    detail = _public_detail(error.detail, error.status_code)
    logger.warning(
        "api_error path={} status={} code={}", _.url.path, error.status_code, detail["code"]
    )
    return JSONResponse(status_code=error.status_code, content={"error": detail})


@app.exception_handler(RequestValidationError)
async def validation_error_handler(_: Request, __: RequestValidationError) -> JSONResponse:
    logger.warning("api_error path={} status=422 code=invalid_request", _.url.path)
    return JSONResponse(status_code=422, content={"error": VOICE_INVALID_REQUEST.detail()})


@app.exception_handler(Exception)
async def unexpected_error_handler(_: Request, error: Exception) -> JSONResponse:
    # Never return exception text: provider responses may contain credentials or conversation content.
    logger.error("agent_unexpected_error exception_type={}", type(error).__name__)
    public = PublicError(
        "internal_error", "The service could not complete that request.", 500, True
    )
    return JSONResponse(status_code=500, content={"error": public.detail()})


def require_service(
    authorization: Annotated[str | None, Header()] = None,
    x_user_id: Annotated[str | None, Header()] = None,
) -> str:
    expected = f"Bearer {get_settings().internal_api_secret}"
    if not authorization or not compare_digest(authorization, expected):
        raise SERVICE_UNAUTHORIZED.as_http_exception()
    if not x_user_id or not 1 <= len(x_user_id) <= 128:
        raise USER_CONTEXT_REQUIRED.as_http_exception()
    return x_user_id


@app.get("/health", operation_id="health")
def health() -> dict[str, str]:
    return {"status": "ok", "service": "riverline-agent"}


@app.get("/v1/workspace", response_model=Workspace, operation_id="get_workspace")
async def workspace(
    request: Request, user_id: Annotated[str, Depends(require_service)]
) -> Workspace:
    finance = getattr(request.app.state, "finance", None)
    if not finance:
        start = datetime.now(ZoneInfo("Asia/Kolkata")).date()
        return Workspace(window_start=start, window_end_exclusive=start + timedelta(days=30))
    snapshot = await finance.workspace(user_id)
    return snapshot.model_copy(update={"voice_available": get_settings().voice_ready})


@app.get("/v1/conversations/latest", operation_id="get_latest_conversation")
async def latest_conversation(
    request: Request, user_id: Annotated[str, Depends(require_service)]
) -> dict | None:
    finance = getattr(request.app.state, "finance", None)
    if not finance:
        return None
    conversations = await finance.conversations(user_id)
    if not conversations:
        return None
    return await finance.conversation(user_id, str(conversations[0]["id"]))


@app.post("/v1/voice/start", response_model=VoiceConnection, operation_id="start_voice")
async def start_voice(
    body: StartVoice,
    request: Request,
    response: Response,
    user_id: Annotated[str, Depends(require_service)],
) -> VoiceConnection:
    response.headers["Cache-Control"] = "no-store"
    return await request.app.state.voice.start(user_id, body.request_id, body.conversation_mode)


@app.get("/v1/voice/{session_id}", response_model=VoiceState, operation_id="get_voice")
async def get_voice(
    session_id: str, request: Request, user_id: Annotated[str, Depends(require_service)]
) -> VoiceState:
    return request.app.state.voice.get(user_id, session_id).state()


@app.post("/v1/voice/{session_id}/end", response_model=VoiceState, operation_id="end_voice")
async def end_voice(
    session_id: str, request: Request, user_id: Annotated[str, Depends(require_service)]
) -> VoiceState:
    return await request.app.state.voice.end(user_id, session_id)