# services/agent/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Riverline Agent API",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Health"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workspace": {
      "get": {
        "summary": "Workspace",
        "operationId": "get_workspace",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workspace"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/conversations/latest": {
      "get": {
        "summary": "Latest Conversation",
        "operationId": "get_latest_conversation",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "type": "object",
                      "additionalProperties": true
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Latest Conversation"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/voice/start": {
      "post": {
        "summary": "Start Voice",
        "operationId": "start_voice",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartVoice"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceConnection"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/voice/{session_id}": {
      "get": {
        "summary": "Get Voice",
        "operationId": "get_voice",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/voice/{session_id}/end": {
      "post": {
        "summary": "End Voice",
        "operationId": "end_voice",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MoneyFact": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "amount_paise": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 9007199254740991.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount Paise"
          },
          "min_amount_paise": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 9007199254740991.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Amount Paise"
          },
          "max_amount_paise": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 9007199254740991.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Amount Paise"
          },
          "usable_amount_paise": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 9007199254740991.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Usable Amount Paise"
          },
          "restricted": {
            "type": "boolean",
            "title": "Restricted",
            "default": false
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "recurring_day_of_month": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 31.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Recurring Day Of Month"
          },
          "timing_note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 140
              },
              {
                "type": "null"
              }
            ],
            "title": "Timing Note"
          },
          "certainty": {
            "type": "string",
            "enum": [
              "confirmed",
              "estimated",
              "uncertain",
              "unknown"
            ],
            "title": "Certainty",
            "default": "unknown"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label"
        ],
        "title": "MoneyFact"
      },
      "PlanEvent": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "change_paise": {
            "type": "integer",
            "title": "Change Paise"
          },
          "balance_paise": {
            "type": "integer",
            "title": "Balance Paise"
          }
        },
        "type": "object",
        "required": [
          "date",
          "label",
          "change_paise",
          "balance_paise"
        ],
        "title": "PlanEvent"
      },
      "PlanSummary": {
        "properties": {
          "dated_income_paise": {
            "type": "integer",
            "title": "Dated Income Paise",
            "default": 0
          },
          "dated_outgoings_paise": {
            "type": "integer",
            "title": "Dated Outgoings Paise",
            "default": 0
          },
          "projected_closing_paise": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Projected Closing Paise"
          },
          "lowest_balance_paise": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lowest Balance Paise"
          },
          "first_shortfall_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Shortfall Date"
          },
          "unplanned_fact_count": {
            "type": "integer",
            "title": "Unplanned Fact Count",
            "default": 0
          },
          "uncertain_income_paise": {
            "type": "integer",
            "title": "Uncertain Income Paise",
            "default": 0
          },
          "uncertain_outgoings_paise": {
            "type": "integer",
            "title": "Uncertain Outgoings Paise",
            "default": 0
          },
          "conditional_closing_paise": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conditional Closing Paise"
          }
        },
        "type": "object",
        "title": "PlanSummary"
      },
      "StartVoice": {
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "conversation_mode": {
            "type": "string",
            "enum": [
              "new",
              "returning"
            ],
            "title": "Conversation Mode",
            "default": "returning"
          }
        },
        "type": "object",
        "required": [
          "request_id"
        ],
        "title": "StartVoice"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VoiceConnection": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "starting",
              "active",
              "ended",
              "failed"
            ],
            "title": "Status"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable",
            "default": false
          },
          "activity": {
            "type": "string",
            "enum": [
              "connecting",
              "thinking",
              "speaking",
              "listening",
              "ended",
              "failed"
            ],
            "title": "Activity",
            "default": "connecting"
          },
          "room_url": {
            "type": "string",
            "title": "Room Url"
          },
          "token": {
            "type": "string",
            "title": "Token"
          },
          "expires_at": {
            "type": "integer",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "status",
          "room_url",
          "token",
          "expires_at"
        ],
        "title": "VoiceConnection"
      },
      "VoiceState": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "starting",
              "active",
              "ended",
              "failed"
            ],
            "title": "Status"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable",
            "default": false
          },
          "activity": {
            "type": "string",
            "enum": [
              "connecting",
              "thinking",
              "speaking",
              "listening",
              "ended",
              "failed"
            ],
            "title": "Activity",
            "default": "connecting"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "status"
        ],
        "title": "VoiceState"
      },
      "Workspace": {
        "properties": {
          "revision": {
            "type": "integer",
            "title": "Revision",
            "default": 0
          },
          "currency": {
            "type": "string",
            "const": "INR",
            "title": "Currency",
            "default": "INR"
          },
          "window_start": {
            "type": "string",
            "format": "date",
            "title": "Window Start"
          },
          "window_end_exclusive": {
            "type": "string",
            "format": "date",
            "title": "Window End Exclusive"
          },
          "opening_cash_paise": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opening Cash Paise"
          },
          "opening_cash": {
            "items": {
              "$ref": "#/components/schemas/MoneyFact"
            },
            "type": "array",
            "title": "Opening Cash"
          },
          "income": {
            "items": {
              "$ref": "#/components/schemas/MoneyFact"
            },
            "type": "array",
            "title": "Income"
          },
          "commitments": {
            "items": {
              "$ref": "#/components/schemas/MoneyFact"
            },
            "type": "array",
            "title": "Commitments"
          },
          "expenses": {
            "items": {
              "$ref": "#/components/schemas/MoneyFact"
            },
            "type": "array",
            "title": "Expenses"
          },
          "plan_status": {
            "type": "string",
            "enum": [
              "not_started",
              "building",
              "ready"
            ],
            "title": "Plan Status",
            "default": "not_started"
          },
          "summary": {
            "$ref": "#/components/schemas/PlanSummary"
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/PlanEvent"
            },
            "type": "array",
            "title": "Timeline"
          },
          "voice_available": {
            "type": "boolean",
            "title": "Voice Available",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "window_start",
          "window_end_exclusive"
        ],
        "title": "Workspace"
      }
    }
  }
}